# rncp_rapport3_preuves/exports/openapi.json
{"openapi":"3.1.0","info":{"title":"MailSavior IA API","version":"1.0.0"},"paths":{"/health":{"get":{"summary":"Health","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/analyze_email":{"post":{"summary":"Analyze Email","operationId":"analyze_email_analyze_email_post","parameters":[{"name":"x-api-key","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"X-Api-Key"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmailRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AnalyzeResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/feedback":{"post":{"summary":"Submit Feedback","operationId":"submit_feedback_feedback_post","parameters":[{"name":"x-api-key","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"X-Api-Key"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FeedbackRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/feedbacks":{"get":{"summary":"Get Feedbacks","operationId":"get_feedbacks_feedbacks_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"AnalyzeResponse":{"properties":{"label":{"type":"string","title":"Label"},"suspicion_score":{"type":"number","title":"Suspicion Score"},"model_version":{"type":"string","title":"Model Version"}},"type":"object","required":["label","suspicion_score","model_version"],"title":"AnalyzeResponse"},"EmailRequest":{"properties":{"text":{"type":"string","maxLength":10000,"minLength":1,"title":"Text"}},"type":"object","required":["text"],"title":"EmailRequest"},"FeedbackRequest":{"properties":{"suspicion_score":{"type":"number","maximum":1.0,"minimum":0.0,"title":"Suspicion Score"},"feedback":{"type":"string","minLength":1,"title":"Feedback"},"predicted_label":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Predicted Label"},"text_length":{"anyOf":[{"type":"integer","minimum":0.0},{"type":"null"}],"title":"Text Length"}},"type":"object","required":["suspicion_score","feedback"],"title":"FeedbackRequest"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}